Correct Task 1

+++ main.h
@@ -22,7 +22,7 @@ typedef struct buf
 int _printf(const char *format, ...);
 int print_char(char c, BUFFER *bp);
 int print_str(char *str, BUFFER *bp);
-int print_dec(int val, BUFFER *bp);
+int print_dec(long int val, BUFFER *bp);
 int print_bin(unsigned int n, BUFFER *bp);
 int print_udec(unsigned int n, BUFFER *bp);
 int print_oct(unsigned int n, BUFFER *bp);
